# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "B2B/B2C E-Commerce API Documentation",
        "description": "API documentation for a hybrid B2B/B2C e-commerce platform for occupational health and safety clothing",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://b2bb2c.mutfakyapim.net",
            "description": "Production API Server"
        },
        {
            "url": "http://127.0.0.1:8000",
            "description": "Development API Server"
        }
    ],
    "paths": {
...
        "/api/v1/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get user orders",
                "description": "Retrieve orders for authenticated user with optional filtering",
                "operationId": "getOrders",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by order status",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "from_date",
                        "in": "query",
                        "description": "Filter orders from date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to_date",
                        "in": "query",
                        "description": "Filter orders to date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page (default: 15)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Orders retrieved successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create new order from cart",
                "description": "Process checkout and create order from user's cart",
                "operationId": "createOrder",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "shipping_address",
                                    "billing_address"
                                ],
                                "properties": {
                                    "shipping_address": {
                                        "type": "object"
                                    },
                                    "billing_address": {
                                        "type": "object"
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "enum": [
                                            "card",
                                            "credit",
                                            "bank_transfer"
                                        ]
                                    },
                                    "notes": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created successfully"
                    },
                    "400": {
                        "description": "Cart is empty or validation failed"
                    },
                    "422": {
                        "description": "Checkout validation failed"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/orders/{order}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get specific order",
                "description": "Retrieve details of a specific order",
                "operationId": "getOrder",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order retrieved successfully"
                    },
                    "403": {
                        "description": "Not authorized to view this order"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/orders/guest-checkout": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Process guest checkout",
                "description": "Create order without user authentication",
                "operationId": "guestCheckout",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cart_data",
                                    "shipping_address",
                                    "billing_address"
                                ],
                                "properties": {
                                    "cart_data": {
                                        "properties": {
                                            "items": {
                                                "type": "array",
                                                "items": {
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "shipping_address": {
                                        "type": "object"
                                    },
                                    "billing_address": {
                                        "type": "object"
                                    },
                                    "customer_email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "enum": [
                                            "card",
                                            "credit",
                                            "bank_transfer"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Guest order created successfully"
                    },
                    "400": {
                        "description": "Cart data is required for guest checkout"
                    }
                }
            }
        },
        "/api/v1/orders/{order}/tracking": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get order tracking information",
                "description": "Retrieve tracking details for an order using order number",
                "operationId": "getOrderTracking",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "Order number",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tracking information retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "order_number": {
                                                    "type": "string"
                                                },
                                                "status": {
                                                    "type": "string"
                                                },
                                                "tracking_number": {
                                                    "type": "string"
                                                },
                                                "shipping_carrier": {
                                                    "type": "string"
                                                },
                                                "estimated_delivery": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/orders/user/summary": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get user order summary",
                "description": "Retrieve order statistics and summary for authenticated user",
                "operationId": "getUserOrderSummary",
                "responses": {
                    "200": {
                        "description": "Order summary retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "total_orders": {
                                                    "type": "integer"
                                                },
                                                "total_spent": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "recent_orders": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "status_counts": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            },
            "0": []
        }
    },
...
}